requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.29.3
pytest==8.3.4
pytest-asyncio==0.25.2
